refactor: Piece 상속 구조 변경 리팩토링
모든 기물이 Piece를 상속 받았었는데,
Pawn/Knight : SpecialPiece 상속
King/Queen/Rook/Bishop : NormalPiece 상속
