+++ src/main/java/chess/piece/Pieces.java
@@ -3,6 +3,13 @@ package chess.piece;
 import chess.board.File;
 import chess.board.Position;
 import chess.board.Rank;
+import chess.piece.normal.Bishop;
+import chess.piece.normal.King;
+import chess.piece.normal.Queen;
+import chess.piece.normal.Rook;
+import chess.piece.special.Knight;
+import chess.piece.special.Pawn;
+
 import java.util.ArrayList;
 import java.util.List;
