+++ src/test/java/chess/piece/PawnTest.java
@@ -7,6 +7,9 @@ import chess.board.Position;
 import chess.board.Rank;
 import java.util.Collections;
 import java.util.List;
+
+import chess.piece.normal.King;
+import chess.piece.special.Pawn;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
