+++ src/test/java/chess/piece/PieceTest.java
@@ -7,6 +7,12 @@ import java.util.stream.Stream;
 import chess.board.File;
 import chess.board.Position;
 import chess.board.Rank;
+import chess.piece.normal.Bishop;
+import chess.piece.normal.King;
+import chess.piece.normal.Queen;
+import chess.piece.normal.Rook;
+import chess.piece.special.Knight;
+import chess.piece.special.Pawn;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
